% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa}
\alias{get_vaa}
\title{NHDPlusV2 Attribute Subset}
\usage{
get_vaa(
  atts = NULL,
  path = get_vaa_path(),
  download = TRUE,
  updated_network = FALSE
)
}
\arguments{
\item{atts}{character The variable names you would like, always includes comid}

\item{path}{character path where the file should be saved. Default is a
persistent system data as retrieved by \link{hydrogeofetch_data_dir}.
Also see: \link{get_vaa_path}}

\item{download}{logical if TRUE, the default, will download VAA table if not
found at path.}

\item{updated_network}{logical default FALSE. If TRUE, updated network attributes
from E2NHD and National Water Model retrieved from
\doi{10.5066/P976XCVT}.}
}
\value{
data.frame containing requested VAA data
}
\description{
Return requested NHDPlusV2 Attributes.

Calling this function downloads the attribute table to \link{get_vaa_path}
if it is not already cached there. The base table is roughly 260 MB and the
updated network table another 170 MB, so this function has no example. To
try it without writing to your user data directory, point
\link{hydrogeofetch_data_dir} at a temporary path first:
\preformatted{
old_dir <- hydrogeofetch_data_dir()
hydrogeofetch_data_dir(file.path(tempdir(), "vaa"))

get_vaa("slope")
get_vaa(c("slope", "lengthkm"))
get_vaa("reachcode", updated_network = TRUE)

hydrogeofetch_data_dir(old_dir)}
}
\details{
The VAA data is a aggregate table of information from the NHDPlusV2
elevslope.dbf(s), PlusFlowlineVAA.dbf(s); and NHDFlowlines. All data
originates from the EPA NHDPlus Homepage
\href{https://www.epa.gov/waterdata/get-nhdplus-national-hydrography-dataset-plus-data}{here}.
To see the location of cached data on your machine use
\code{\link{get_vaa_path}}.
To view aggregate data and documentation, see
\href{https://www.hydroshare.org/resource/6092c8a62fac45be97a09bfd0b0bf726/}{here}
}
