% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa_names}
\alias{get_vaa_names}
\title{Available NHDPlusV2 Attributes}
\usage{
get_vaa_names(updated_network = FALSE)
}
\arguments{
\item{updated_network}{logical default FALSE. If TRUE, updated network attributes
from E2NHD and National Water Model retrieved from
\doi{10.5066/P976XCVT}.}
}
\value{
character vector
}
\description{
Find variables available from the NHDPlusV2 attribute data.frame

Calling this function downloads the attribute table to \link{get_vaa_path}
if it is not already cached there. The table is roughly 260 MB, so this
function has no example. 

\preformatted{
old_dir <- hydrogeofetch_data_dir()
hydrogeofetch_data_dir(file.path(tempdir(), "vaa"))

get_vaa_names()

hydrogeofetch_data_dir(old_dir)}
}
\details{
The VAA data is a aggregate table of information from the NHDPlusV2
elevslope.dbf(s), PlusFlowlineVAA.dbf(s); and NHDFlowlines. All data
originates from the EPA NHDPlus Homepage
\href{https://www.epa.gov/waterdata/get-nhdplus-national-hydrography-dataset-plus-data}{here}.
To see the location of cached data on your machine use
\code{\link{get_vaa_path}}.
To view aggregate data and documentation, see
\href{https://www.hydroshare.org/resource/6092c8a62fac45be97a09bfd0b0bf726/}{here}
}
